RFC: Allow passing ACPI tables to bootefi
At present it seems possible to pass ACPI tables using bootefi but
in practice an error is produced about not beling allowed to pass
devicetree as well.

Add some workarounds for this. Presumably I am missing something and
this should already work. Perhaps fdt should be set to 0?

Not to be applied.

Signed-off-by: Simon Glass <sjg@chromium.org>